SQL encoding failed with bbox expresions which had a null left or right expersion.
These are normaly interprited as being the default geometry, to avoid this problem a name for the default geometry can be set and this will be used in the event of a null geometry expression being found

Needs some more work, can the default geometry be determined?  Should we always pass in a Schema so that all filter encoders can do this trick if they need to?


git-svn-id: http://svn.osgeo.org/geotools/trunk@1873 e5c1c795-43da-0310-a71f-fac65c449510