Commit the change for the second time. :(
changed the chopping of .shp to operate on the toString version of the URL
instead of just the getFile method, and hence avoid the relative URL
constructor. The reason is that resources embedded in jars that are loaded
via the classloader start with "jar:file", but the relative URL method was
changing this to just plain "file", which seems like a bug in the URL class
to me, but...  Once this happens the URL no longer points to the same
location and it will no longer be able to find the shp file. Using the
toString seems to work and does not seem to have any downside


git-svn-id: http://svn.osgeo.org/geotools/trunk@2960 e5c1c795-43da-0310-a71f-fac65c449510